refactor: Move to a stanadlone function
